@@ -1,36 +1,26 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
-import Home from './pages/Home/index'
-import Error from './components/Error/index'
-import About from './pages/About/index'
-import Housing from './pages/Housing/index'
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+import Home from "./pages/Home/index";
+import Error from "./components/Error/index";
+import About from "./pages/About/index";
+import Housing from "./pages/Housing/index";
 import "./styles/global.scss";
-import Header from './components/Header/index'
-import Footer from './components/Footer/index'
+import Footer from "./components/Footer/index";
+import Header from "./components/Header/index";
 
-
-
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
   <React.StrictMode>
     <Router>
       <Header />
-      <Switch>
-        <Route exact path="/">
-          <Home />
-        </Route>
-        <Route path="/About">
-          <About />
-        </Route>
-        <Route path="/Housing">
-          <Housing />
-        </Route>
-        <Route path="*">
-          <Error />
-        </Route>
-      </Switch>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/about" element={<About />} />
+        <Route path="/logement/:id" element={<Housing />} />
+        <Route path="*" element={<Error />} />
+      </Routes>
       <Footer />
     </Router>
-  </React.StrictMode>,
-  document.getElementById('root')
-)
+  </React.StrictMode>
+);
